Commit local
The initial simulated_observation.py is modified to simulated_obs.py. This defines a function sim_obs which defines the 2d box to simulate the observation, takes x and y shift in pixels and provides a simulated frame.

The second function is jitter_sim_stel.py.  This has a function jit_sim, which takes a sun center info from a file path and dethers the simulated observation accordingly. It can save the output of every frame as .png and .fits with user provided flags.